requirements.txt
moviepy
edge-tts
google-generativeai
Pillow
openai
requests
